Re-imported data.

+++ PD/Models/AppViewModels/Faculty/FacultySalaryViewModel.cs
@@ -7,7 +7,7 @@ namespace PD.Models.AppViewModels.Faculty
 {
     public class FacultySalaryViewModel : SalaryViewModel
     {
-        public double MeritDecision { get; set; }
+        public decimal MeritDecision { get; set; }
         public string MeritReason { get; set; }
         public decimal ContractSettlement { get; set; }
         public decimal Merit { get; set; }
